requirements.txt
pytest
psycopg2-binary
allure-pytest